+++ src/core/component.js
@@ -1,4 +1,11 @@
-const { markdown } = require('markdown');
+const marked = require('marked');
+marked.setOptions({
+  breaks: false,
+  smartLists: false,
+  smartypants: false,
+  xhtml: true
+});
+
 const { validator, flatten } = require('./utilities');
 const _ = require('lodash');
 
@@ -120,20 +127,11 @@ class Component {
 
     return this;
   }
-  get notesMdTree(){
-    if(this.notes){
-      return markdown.parse(this.notes);
-    }else{
-      return undefined;
-    }
-  }
   get notesHTML() {
-    if(this.notes){
-      let HTMLTree = markdown.toHTMLTree(this.notesMdTree);
-      return markdown.renderJsonML(HTMLTree);
-    }else{
+    if (this.notes === undefined) {
       return undefined;
     }
+    return marked(this.notes);
   }
   static isValid(q, logger){
     let ind = q.space ? `${q.space}::${q.id}` : q.id;